// src/model/schedule-manage-info.d.ts
export interface ScheduleManageInfo {
  codeScheduleManage: string;//项目单元编号
  nameScheduleManage: string;//项目单元名称
  typeScheduleManage: string;//项目性质

  progressOverviewScheduleManage: string;//进展概述
  completePercentageScheduleManage: string;//完工百分比
  designCompletedDateScheduleManage: string;//设计完成时间
  drawCompletedDateScheduleManage: string;//施工图完成时间
  planCompletionDateScheduleManage: string;//计划完工时间
  actualStartDateScheduleManage: string;//实际开工时间
  planStartDateScheduleManage: string;//计划开工时间
  practicalCompletionDateScheduleManage: string;//实际完工时间
  finalAcceptanceDateScheduleManage: string;//竣工验收时间
  preTransferDateScheduleManage: string;//预达转资时间
  auditReportDateScheduleManage: string;//审计报告日期
  applyDateScheduleManage: string;//填报日期
  applyUserScheduleManage: string;//填报人
}